web_ui/requirements.txt
flask
happybase
kafka-python
